test(#781): split use_font_blocks tests to satisfy max-lines
The new use_font_blocks tests pushed ProjectConfigParser.test.ts over
the 500-line per-file limit (ESLint max-lines, error). Move them into
their own ProjectConfigParser.useFontBlocks.test.ts file.
